Send replaced branch for pr
This fixes an issue with CI status in SCM-Manager for
Jenkins multibranch pipelines.
If such a pipeline fails for a branch for whatever
reason (eg. due to a flappy test), this result is stored
in SCM-Manager for the revision. If later on a pull
request is created for this branch, this failed build
is considered as a failed status for the pull request,
even though the pull request build itself might be
successful.

To fix this, this plugin computes the name of the
branch build for a pull request build and sends this
name as 'replaces' within the status request for
SCM-Manager. This 'replaces' field will be evaluated
with the CI plugin version 2.6.0 or later in SCM-Manager.