[WIP] Replace Repository Combobox to InputBox
Previously, there has been a comboBox in the config.jelly file for the
repository entry. Since this comes the caveat of not being able to paste
text onto it and this is explicitly not desired to be the case, it has
been changed into a standard text input with an autocomplete feature.
Note that this autocomplete feature is not yet functional and requires
adaptions in the class ScmManagerApi.

The autocomplete method cannot take the serverUrl and credentialsId as
query parameters, so they need to be stored by the servlet within the
class in order for the autocomplete field to be functional.

Furthermore, the SCM type acronym has been removed from shown text.
Other minor additions include adding an mvnDebug wrapper and a
mise-en-place file.

Signed-off-by: Till-André Diegeler <till-andre.diegeler@cloudogu.com>