Fix + extend unit tests, POM fixes
For the source descriptor, some new form messages were introduced and fixes done to the existing unit tests, which were also extended. In this revision, a caching system was tested, but this isn't reflected in the final changeset and this not visible.

Furthermore, some texts have been
improved. A JUnit build fix has been done in the pom.xml