Integration of Custom Properties for Multibranch Pipelines with SCM-Manager Source (#71)
Added the fetching of the custom properties of the corresponding SCM-Manager Repository.
The custom properties will be provided as environment variable for every build.
Each environment variable will be named after a prefix and the corresponding key of the custom property. For example the custom property with the key "lang", will be provided as the environment variable "SCM_CUSTOM_PROP_lang"

---------

Co-authored-by: Viktor <viktor.egorov@triology.de>